work to make tanstack tables more dynamic
